Move `completion_mechanism` to its own file
